Rename module/etc to eyes as well
