+++ modules/eyes/errors.go
